refactor(hub): I18nStrategy + NO_I18N + dict predicates move to the kernel port (#623)
