+++ packages/hub/src/kernel/collection-config.ts
@@ -31,7 +31,7 @@ import type { LedgerStore } from '../with-commit/history/ledger/index.js'
 import type { CrdtMode } from '../with-commit/crdt/crdt.js'
 import { NO_CRDT, type CrdtStrategy } from '../with-commit/crdt/strategy.js'
 import { NO_HISTORY, type HistoryStrategy } from '../with-commit/history/strategy.js'
-import { NO_I18N, type I18nStrategy } from '../shape/via-i18n/strategy.js'
+import { NO_I18N, type I18nStrategy } from '../port/with/i18n-strategy.js'
 import { NO_SYNC, type SyncStrategy } from '../with-party/team/sync-strategy.js'
 import { NO_BLOBS, type BlobStrategy } from '../with-shape/blobs/strategy.js'
 import { NO_AGGREGATE, type AggregateStrategy } from '../with-lookup/aggregate/strategy.js'
