+++ packages/hub/src/kernel/collection.ts
@@ -11,12 +11,12 @@ import type { I18nTextDescriptor } from '../shape/via-i18n/core.js'
 import { stripI18nFilled } from '../shape/via-i18n/core.js'
 import { getAtPath, setAtPathInPlace } from './paths.js'
 import type { DictKeyDescriptor, StaticDictDescriptor, DictionaryHandle } from '../shape/via-i18n/dictionary.js'
-import { isStaticDictDescriptor } from '../shape/via-i18n/dictionary.js'
+import { isStaticDictDescriptor } from '../port/with/i18n-strategy.js'
 import { ViaPipeline } from './via-pipeline.js'
 import { viaBinder, type ViaDescriptor } from './via.js'
 import type { ComputedFields } from '../with-formula/computed/index.js'
 import { enforceClassifiedWrite } from '../with-shape/classified/write.js'
-import type { I18nStrategy } from '../shape/via-i18n/strategy.js'
+import type { I18nStrategy } from '../port/with/i18n-strategy.js'
 import { resolvePolicy } from '../shape/via-i18n/policy.js'
 import {
   isTombstone,
