+++ scripts/check-architecture.mjs
@@ -1168,7 +1168,6 @@ const PRE_EXISTING_SPINE_SERVICE_IMPORTS = new Map([
     '../with-shape/classified/strategy.js',
     '../shape/via-i18n/core.js',
     '../shape/via-i18n/dictionary.js',
-    '../shape/via-i18n/strategy.js',
     '../with-shape/introspection/field-meta.js',
     '../with-shape/introspection/meta.js',
     '../with-shape/schema-update/fence-controller.js',
@@ -1230,7 +1229,6 @@ const PRE_EXISTING_SPINE_SERVICE_IMPORTS = new Map([
     '../shape/via-i18n/core.js',
     '../shape/via-i18n/dictionary.js',
     '../shape/via-i18n/policy.js',
-    '../shape/via-i18n/strategy.js',
     '../with-shape/introspection/describe.js',
     '../with-shape/introspection/field-meta.js',
     '../with-shape/introspection/meta.js',
@@ -1375,7 +1373,6 @@ const PRE_EXISTING_SPINE_SERVICE_IMPORTS = new Map([
     '../with-shape/classified/strategy.js',
     '../shape/via-i18n/core.js',
     '../shape/via-i18n/dictionary.js',
-    '../shape/via-i18n/strategy.js',
     '../with-shape/introspection/field-meta.js',
     '../with-shape/introspection/meta.js',
     '../with-shape/introspection/types.js',
